+++ docassemble_base/docassemble/base/util.py
@@ -11,7 +11,7 @@ from celery import chord
 from docassemble.base.logger import logmessage
 from docassemble.base.error import DAError
 from docassemble.base.functions import alpha, roman, item_label, comma_and_list, get_language, set_language, get_dialect, set_country, get_country, word, comma_list, ordinal, ordinal_number, need, nice_number, quantity_noun, possessify, verb_past, verb_present, noun_plural, noun_singular, space_to_underscore, force_ask, force_gather, period_list, name_suffix, currency_symbol, currency, indefinite_article, nodoublequote, capitalize, title_case, url_of, do_you, did_you, does_a_b, did_a_b, were_you, was_a_b, have_you, has_a_b, your, her, his, is_word, get_locale, set_locale, process_action, url_action, get_info, set_info, get_config, prevent_going_back, qr_code, action_menu_item, from_b64_json, defined, value, message, response, json_response, command, single_paragraph, quote_paragraphs, location_returned, location_known, user_lat_lon, interview_url, interview_url_action, interview_url_as_qr, interview_url_action_as_qr, interview_email, get_emails, this_thread, static_image, action_arguments, action_argument, language_functions, language_function_constructor, get_default_timezone, user_logged_in, interface, user_privileges, user_has_privilege, user_info, task_performed, task_not_yet_performed, mark_task_as_performed, times_task_performed, set_task_counter, background_action, background_response, background_response_action, us, set_live_help_status, chat_partners_available, phone_number_in_e164, phone_number_is_valid, countries_list, country_name, write_record, read_records, delete_record, variables_as_json, all_variables, server, language_from_browser, device, plain, bold, italic, states_list, state_name, subdivision_type, indent, raw, fix_punctuation, set_progress, get_progress, referring_url, undefine, dispatch, yesno, noyes, split, showif, showifdef, phone_number_part
-from docassemble.base.core import DAObject, DAList, DADict, DASet, DAFile, DAFileCollection, DAFileList, DAEmail, DAEmailRecipient, DAEmailRecipientList, DATemplate, selections, objects_from_file
+from docassemble.base.core import DAObject, DAList, DADict, DASet, DAFile, DAFileCollection, DAStaticFile, DAFileList, DAEmail, DAEmailRecipient, DAEmailRecipientList, DATemplate, selections, objects_from_file
 from decimal import Decimal
 import sys
 #sys.stderr.write("importing async mail now from util\n")
@@ -34,7 +34,7 @@ import shutil
 import subprocess
 from bs4 import BeautifulSoup
 
-__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'force_gather', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'LatitudeLongitude', 'RoleChangeTracker', 'Name', 'IndividualName', 'Address', 'City', 'Event', 'Person', 'Thing', 'Individual', 'ChildList', 'FinancialList', 'PeriodicFinancialList', 'Income', 'Asset', 'Expense', 'Value', 'PeriodicValue', 'OfficeList', 'Organization', 'objects_from_file', 'send_email', 'send_sms', 'map_of', 'selections', 'DAObject', 'DAList', 'DADict', 'DASet', 'DAFile', 'DAFileCollection', 'DAFileList', 'DAEmail', 'DAEmailRecipient', 'DAEmailRecipientList', 'DATemplate', 'last_access_time', 'last_access_delta', 'last_access_days', 'last_access_hours', 'last_access_minutes', 'returning_user', 'action_arguments', 'action_argument', 'timezone_list', 'as_datetime', 'current_datetime', 'date_difference', 'date_interval', 'year_of', 'month_of', 'day_of', 'dow_of', 'format_date', 'format_time', 'today', 'get_default_timezone', 'user_logged_in', 'interface', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'DARedis', 'MachineLearningEntry', 'SimpleTextMachineLearner', 'SVMMachineLearner', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'ocr_file', 'ocr_file_in_background', 'read_qr', 'get_sms_session', 'initiate_sms_session', 'terminate_sms_session', 'language_from_browser', 'device', 'interview_email', 'get_emails', 'plain', 'bold', 'italic', 'path_and_mimetype', 'states_list', 'state_name', 'subdivision_type', 'indent', 'raw', 'fix_punctuation', 'set_progress', 'get_progress', 'referring_url', 'run_python_module', 'undefine', 'dispatch', 'yesno', 'noyes', 'split', 'showif', 'showifdef', 'phone_number_part']
+__all__ = ['alpha', 'roman', 'item_label', 'ordinal', 'ordinal_number', 'comma_list', 'word', 'get_language', 'set_language', 'get_dialect', 'set_country', 'get_country', 'get_locale', 'set_locale', 'comma_and_list', 'need', 'nice_number', 'quantity_noun', 'currency_symbol', 'verb_past', 'verb_present', 'noun_plural', 'noun_singular', 'indefinite_article', 'capitalize', 'space_to_underscore', 'force_ask', 'force_gather', 'period_list', 'name_suffix', 'currency', 'static_image', 'title_case', 'url_of', 'process_action', 'url_action', 'get_info', 'set_info', 'get_config', 'prevent_going_back', 'qr_code', 'action_menu_item', 'from_b64_json', 'defined', 'value', 'message', 'response', 'json_response', 'command', 'single_paragraph', 'quote_paragraphs', 'location_returned', 'location_known', 'user_lat_lon', 'interview_url', 'interview_url_action', 'interview_url_as_qr', 'interview_url_action_as_qr', 'LatitudeLongitude', 'RoleChangeTracker', 'Name', 'IndividualName', 'Address', 'City', 'Event', 'Person', 'Thing', 'Individual', 'ChildList', 'FinancialList', 'PeriodicFinancialList', 'Income', 'Asset', 'Expense', 'Value', 'PeriodicValue', 'OfficeList', 'Organization', 'objects_from_file', 'send_email', 'send_sms', 'map_of', 'selections', 'DAObject', 'DAList', 'DADict', 'DASet', 'DAFile', 'DAFileCollection', 'DAFileList', 'DAStaticFile', 'DAEmail', 'DAEmailRecipient', 'DAEmailRecipientList', 'DATemplate', 'last_access_time', 'last_access_delta', 'last_access_days', 'last_access_hours', 'last_access_minutes', 'returning_user', 'action_arguments', 'action_argument', 'timezone_list', 'as_datetime', 'current_datetime', 'date_difference', 'date_interval', 'year_of', 'month_of', 'day_of', 'dow_of', 'format_date', 'format_time', 'today', 'get_default_timezone', 'user_logged_in', 'interface', 'user_privileges', 'user_has_privilege', 'user_info', 'task_performed', 'task_not_yet_performed', 'mark_task_as_performed', 'times_task_performed', 'set_task_counter', 'background_action', 'background_response', 'background_response_action', 'us', 'DARedis', 'MachineLearningEntry', 'SimpleTextMachineLearner', 'SVMMachineLearner', 'set_live_help_status', 'chat_partners_available', 'phone_number_in_e164', 'phone_number_is_valid', 'countries_list', 'country_name', 'write_record', 'read_records', 'delete_record', 'variables_as_json', 'all_variables', 'ocr_file', 'ocr_file_in_background', 'read_qr', 'get_sms_session', 'initiate_sms_session', 'terminate_sms_session', 'language_from_browser', 'device', 'interview_email', 'get_emails', 'plain', 'bold', 'italic', 'path_and_mimetype', 'states_list', 'state_name', 'subdivision_type', 'indent', 'raw', 'fix_punctuation', 'set_progress', 'get_progress', 'referring_url', 'run_python_module', 'undefine', 'dispatch', 'yesno', 'noyes', 'split', 'showif', 'showifdef', 'phone_number_part']
 
 #knn_machine_learner = DummyObject
 
@@ -1134,17 +1134,7 @@ def send_sms(to=None, body=None, template=None, task=None, attachments=None, con
         body = BeautifulSoup(body_html, "html.parser").get_text('\n')
     if body is None:
         body = word("blank message")
-    if this_thread.current_info.get('url_root', None) is not None:
-        url_start = re.sub(r'/$', r'', this_thread.current_info['url_root'])
-    else:
-        url_start = get_config('url root')
-        if url_start is None:
-            url_start = 'http://localhost'
-        url_start = re.sub(r'/$', r'', url_start)
-        root = get_config('root')
-        if root is None:
-            root = '/'
-        url_start += root
+    url_start = docassemble.base.functions.get_url_start()
     success = True
     media = list()
     for attachment in attachments:
@@ -1164,6 +1154,8 @@ def send_sms(to=None, body=None, template=None, task=None, attachments=None, con
                 success = False
         elif type(attachment) is DAFile:
             attachment_list.append(attachment)
+        elif type(attachment) is DAStaticFile:
+            attachment_list.append(attachment)
         elif type(attachment) is DAFileList:
             attachment_list.extend(attachment.elements)
         elif type(attachment) in [str, unicode] and re.search(r'^https?://', attachment):
@@ -1174,8 +1166,10 @@ def send_sms(to=None, body=None, template=None, task=None, attachments=None, con
         if success:
             for the_attachment in attachment_list:
                 if type(the_attachment) is DAFile and the_attachment.ok:
-                    url = url_start + docassemble.base.filter.url_for('serve_stored_file', uid=this_thread.current_info['session'], number=the_attachment.number, filename=the_attachment.filename, extension=the_attachment.extension)
-                    media.append(url)
+                    #url = url_start + server.url_for('serve_stored_file', uid=this_thread.current_info['session'], number=the_attachment.number, filename=the_attachment.filename, extension=the_attachment.extension)
+                    media.append(the_attachment.url_for(_external=True))
+                if type(the_attachment) is DAStaticFile:
+                    media.append(the_attachment.url_for(_external=True))
                 elif type(the_attachment) in [str, unicode]:
                     media.append(the_attachment)
     if len(media) > 10:
@@ -1202,6 +1196,8 @@ def send_email(to=None, sender=None, cc=None, bcc=None, body=None, html=None, su
     """Sends an e-mail and returns whether sending the e-mail was successful."""
     if attachments is None:
         attachments = []
+    if not hasattr(attachments, '__iter__'):
+        attachments = [attachments]
     from flask_mail import Message
     if type(to) is not list:
         to = [to]
@@ -1236,6 +1232,8 @@ def send_email(to=None, sender=None, cc=None, bcc=None, body=None, html=None, su
                 success = False
         elif type(attachment) is DAFile:
             attachment_list.append(attachment)
+        elif type(attachment) is DAStaticFile:
+            attachment_list.append(attachment)
         elif type(attachment) is DAFileList:
             attachment_list.extend(attachment.elements)
         elif type(attachment) in (str, unicode):
@@ -1254,6 +1252,13 @@ def send_email(to=None, sender=None, cc=None, bcc=None, body=None, html=None, su
             success = False
         if success:
             for the_attachment in attachment_list:
+                if isinstance(the_attachment, DAStaticFile):
+                    the_path = the_attachment.path()
+                    with open(the_path, 'rb') as fp:
+                        the_basename = os.path.basename(the_path)
+                        extension, mimetype = server.get_ext_and_mimetype(the_basename)
+                        msg.attach(attachment_name(the_basename, filenames_used), mimetype, fp.read())
+                    continue
                 if the_attachment.ok:
                     if the_attachment.has_specific_filename:
                         file_info = server.file_finder(str(the_attachment.number), filename=the_attachment.filename)
@@ -1494,10 +1499,18 @@ def read_qr(image_file, f=None, l=None, x=None, y=None, W=None, H=None):
         qr = qrtools.QR()
         qr.decode(page)
         codes.append(qr.data)
-    return codes        
+    return codes
 
 def path_and_mimetype(file_ref):
     """Returns a path and the MIME type of a file"""
+    if isinstance(file_ref, DAFileList) and len(file_ref.elements) > 0:
+        file_ref = file_ref.elements[0]
+    elif isinstance(file_ref, DAFileCollection):
+        file_ref = file_ref._first_file()
+    elif isinstance(file_ref, DAStaticFile):
+        path = file_ref.path()
+        extension, mimetype = server.get_ext_and_mimetype(file_ref.filename)
+        return path, mimetype
     if isinstance(file_ref, DAFile):
         if hasattr(file_ref, 'mimetype'):
             mime_type = file_ref.mimetype