+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.1.75]
+### Added
+- `inline width` field modifier for embedded fields.
+### Fixed
+- `DAFile` initialize function now handles keyword arguments as 
+  advertised.
+
 ## [0.1.74] - 2017-10-15
 ### Added
 - `data` block.